create a service folder for email services make sendWelcome mal resueable [#162727551]
