+++ controllers/index.js
@@ -1,4 +1,2 @@
 export { default as Users } from './user';
 export { default as Profile } from './profile';
-export { default as activateAccount } from './activateAccount';
-export { default as sendMail } from './emailService';
