+++ routes/index.js
@@ -1,4 +1,3 @@
 export { default as authRoutes } from './authRoutes';
 export { default as userRoutes } from './userRoutes';
 export { default as profileRoutes } from './profileRoutes';
-export { default as activateRoute } from './activateRoute';
